update comments to use fontawesome instead of material icons, remove material icons from cdn import, fix bug on post page about edit post. add confirm box on edit, delete of post and comment. update header meta tags to enable social cards

+++ client/src/components/comment/CommentItem.jsx
@@ -1,5 +1,6 @@
 import React, { useContext, Fragment } from 'react'
 import Moment from 'react-moment'
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 
 // Import Context
 import { CommentContext } from '@/context/comment/CommentContext'
@@ -20,8 +21,11 @@ const CommentItem = ({ comment }) => {
       setAlert('Please login')
       return
     } else if (user._id === comment.author || user.admin) {
-      deleteComment(comment._id)
-      setAlert('Successfully deleted a comment')
+      if (window.confirm('Are you sure you want to delete this comment?')) {
+        deleteComment(comment._id)
+        setAlert('Successfully deleted a comment')
+      }
+      return
     } else {
       setAlert('Sorry, You are not authorized to do so')
       return
@@ -44,12 +48,17 @@ const CommentItem = ({ comment }) => {
               : ' hidden')
           }
         >
-          <i
+          {/* <i
             className='material-icons text-gray-500 hover:text-gray-400 cursor-pointer'
             onClick={handleDelete}
           >
             delete
-          </i>
+          </i> */}
+          <FontAwesomeIcon
+            icon='fa-solid fa-trash'
+            className='text-gray-500 text-lg hover:text-gray-400 cursor-pointer'
+            onClick={handleDelete}
+          />
         </div>
       </div>
     </Fragment>
